@@ -17,10 +17,13 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.1",
-        "symfony/dependency-injection": "^5.4|^6.1",
-        "symfony/http-kernel": "^5.4|^6.1",
-        "halloverden/symfony-entity-utils-bundle": "^3.0",
+        "php": ">=8.2",
+        "symfony/dependency-injection": "^6.4|^7.1",
+        "symfony/http-kernel": "^6.4|^7.1",
+        "halloverden/symfony-entity-utils-bundle": "^4.3",
         "doctrine/doctrine-bundle": "^2.5"
+    },
+    "require-dev": {
+        "rector/rector": "^1.2"
     }
 }
