@@ -2,10 +2,6 @@
 
 namespace HalloVerden\FeatureFlagBundle\Exception;
 
-/**
- * Class UnableToBuildFeatureFlagException
- *
- * @package HalloVerden\FeatureFlagBundle\Exception
- */
-class UnableToBuildFeatureFlagException extends \Exception {
+
+final class UnableToBuildFeatureFlagException extends \Exception {
 }
