@@ -4,10 +4,6 @@ namespace HalloVerden\FeatureFlagBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-/**
- * Class HalloVerdenFeatureFlagBundle
- *
- * @package HalloVerden\FeatureFlagBundle
- */
-class HalloVerdenFeatureFlagBundle extends Bundle {
+
+final class HalloVerdenFeatureFlagBundle extends Bundle {
 }
